requirements.txt
streamlit
openai
python-dotenv
torch
transformers
trafilatura
googlesearch-python
bert-extractive-summarizer
scikit-learn
numpy